refactor(headless): split Select into per-file parts for tree-shaking
