+++ packages/headless/src/primitives/select/parts.ts
@@ -1,0 +1,8 @@
+export { type SelectItem, type SelectProps, SelectRoot as Root } from './select-root';
+export { type SelectTriggerProps, SelectTrigger as Trigger } from './select-trigger';
+export { type SelectValueProps, SelectValue as Value } from './select-value';
+export { type SelectPortalProps, SelectPortal as Portal } from './select-portal';
+export { type SelectPositionerProps, SelectPositioner as Positioner } from './select-positioner';
+export { type SelectPopupProps, SelectPopup as Popup } from './select-popup';
+export { type SelectOptionProps, SelectOption as Option } from './select-option';
+export { type SelectArrowProps, SelectArrow as Arrow } from './select-arrow';
